Profalux covers: switch to using cluster 0x102, add MAI-ZTP20C remote
This change references remote model MAI-ZTP20C.
More importantly, it switches the control of covers to using the dedicated ClosuresWindowCovering (0x102) cluster.
Sniffing Zigbee traffic of my remotes show that the remotes use this cluster to control the covers, and not 0x8 as currently implemented.
The compliance document for the Profalux motors at https://csa-iot.org/csa_product/ac-motor-for-window-coverings/ also lists that cluster and does NOT list 0x6/0x8.

So I think the correct way to control these covers is 0x102, and certainly it's the way that seems to work with my shutters (all bought in 2023).

This change is likely to break existing setups if older revisions of these motors did not expose 0x102.

0x102 is on endpoint 2, while 0x6/0x8 are on endpoint 1.
An ideal implementation would use 0x102 when it exists, and 0x8 the rest of the time, but I don't know how to implement that.

This change needs testing by other Profalux users!